Fix Array encoding when having values of different encode width. Add a test to validate properies (get/set) of all framing types. Remove Multiple type as it is not a real type.

+++ src/Framing/Transfer.cs
@@ -70,7 +70,7 @@ namespace Amqp.Framing
         public ReceiverSettleMode RcvSettleMode
         {
             get { return this.Fields[6] == null ? ReceiverSettleMode.First : (ReceiverSettleMode)this.Fields[6]; }
-            set { this.Fields[6] = value; }
+            set { this.Fields[6] = (byte)value; }
         }
 
         public DeliveryState State
